+++ src/renderer/components/pr-detail/index.ts
@@ -1,0 +1,7 @@
+export { PRHeader } from './PRHeader';
+export { PRTabs } from './PRTabs';
+export { FileTree } from './FileTree';
+export { MergeConfirmDialog } from './MergeConfirmDialog';
+export { RequestChangesDialog } from './RequestChangesDialog';
+export { CheckoutDropdown } from './CheckoutDropdown';
+export { usePRNavigation } from './usePRNavigation';
